flask-legacy/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
cx_Oracle==8.3.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
Werkzeug==2.3.7
